+++ src/types/customer.ts
@@ -1,16 +1,31 @@
-import { OnvoAddressT } from './address'
+import { OnvoAddressI } from './address'
+import { OmitId } from './generics'
 
 export type OnvoShipping = {
-  address: OnvoAddressT
+  address: OnvoAddressI
   name: string
   phone?: string
 }
 
 export interface OnvoCustomerI {
-  address?: OnvoAddressT
-  description?: string
-  email: string
-  name: string
-  phone?: string
-  shipping?: OnvoShipping
+  id: string
+  address: OnvoAddressI | null
+  amountSpent: number
+  description: string | null
+  email: string | null
+  lastTransactionAt: string
+  mode: 'test' | 'live'
+  name: string | null
+  phone: string | null
+  shipping: OnvoShipping | null
+  transactionsCount: number
+  createdAt: string
+  updatedAt: string
 }
+
+export type InsertOnvoCustomerI = Partial<
+  Omit<
+    OmitId<OnvoCustomerI>,
+    'amountSpent' | 'transactionsCount' | 'createdAt' | 'updatedAt' | 'mode'
+  >
+>